Changement du nom de la classe et début des commentaires de la classe
